Move AC export and reporting to Tools
