+++ apps/web/src/app/tools/reports/pages/organizations/unsecured-websites-report.component.ts
@@ -2,57 +2,53 @@
 // @ts-strict-ignore
 import { Component, OnInit } from "@angular/core";
 import { ActivatedRoute } from "@angular/router";
-import { firstValueFrom } from "rxjs";
+import { firstValueFrom, map } from "rxjs";
 
+import { CollectionService } from "@bitwarden/admin-console/common";
 import { ModalService } from "@bitwarden/angular/services/modal.service";
 import {
   getOrganizationById,
   OrganizationService,
 } from "@bitwarden/common/admin-console/abstractions/organization/organization.service.abstraction";
 import { AccountService } from "@bitwarden/common/auth/abstractions/account.service";
-import { getUserId } from "@bitwarden/common/auth/services/account.service";
 import { I18nService } from "@bitwarden/common/platform/abstractions/i18n.service";
-import { PasswordStrengthServiceAbstraction } from "@bitwarden/common/tools/password-strength";
 import { CipherService } from "@bitwarden/common/vault/abstractions/cipher.service";
 import { SyncService } from "@bitwarden/common/vault/abstractions/sync/sync.service.abstraction";
-import { Cipher } from "@bitwarden/common/vault/models/domain/cipher";
 import { CipherView } from "@bitwarden/common/vault/models/view/cipher.view";
 import { PasswordRepromptService } from "@bitwarden/vault";
 
 // eslint-disable-next-line no-restricted-imports
-import { WeakPasswordsReportComponent as BaseWeakPasswordsReportComponent } from "../../../tools/reports/pages/weak-passwords-report.component";
+import { UnsecuredWebsitesReportComponent as BaseUnsecuredWebsitesReportComponent } from "../unsecured-websites-report.component";
 
 @Component({
-  selector: "app-weak-passwords-report",
-  templateUrl: "../../../tools/reports/pages/weak-passwords-report.component.html",
+  selector: "app-unsecured-websites-report",
+  templateUrl: "../unsecured-websites-report.component.html",
 })
 // eslint-disable-next-line rxjs-angular/prefer-takeuntil
-export class WeakPasswordsReportComponent
-  extends BaseWeakPasswordsReportComponent
+export class UnsecuredWebsitesReportComponent
+  extends BaseUnsecuredWebsitesReportComponent
   implements OnInit
 {
-  manageableCiphers: Cipher[];
-
   constructor(
     cipherService: CipherService,
-    passwordStrengthService: PasswordStrengthServiceAbstraction,
     modalService: ModalService,
     private route: ActivatedRoute,
     organizationService: OrganizationService,
+    protected accountService: AccountService,
     passwordRepromptService: PasswordRepromptService,
     i18nService: I18nService,
     syncService: SyncService,
-    protected accountService: AccountService,
+    collectionService: CollectionService,
   ) {
     super(
       cipherService,
-      passwordStrengthService,
       organizationService,
       accountService,
       modalService,
       passwordRepromptService,
       i18nService,
       syncService,
+      collectionService,
     );
   }
 
@@ -60,23 +56,19 @@ export class WeakPasswordsReportComponent
     this.isAdminConsoleActive = true;
     // eslint-disable-next-line rxjs-angular/prefer-takeuntil, rxjs/no-async-subscribe
     this.route.parent.parent.params.subscribe(async (params) => {
-      const userId = await firstValueFrom(this.accountService.activeAccount$.pipe(getUserId));
-
+      const userId = await firstValueFrom(
+        this.accountService.activeAccount$.pipe(map((a) => a?.id)),
+      );
       this.organization = await firstValueFrom(
         this.organizationService
           .organizations$(userId)
           .pipe(getOrganizationById(params.organizationId)),
       );
-      this.manageableCiphers = await this.cipherService.getAll(userId);
       await super.ngOnInit();
     });
   }
 
   getAllCiphers(): Promise<CipherView[]> {
     return this.cipherService.getAllFromApiForOrganization(this.organization.id);
-  }
-
-  canManageCipher(c: CipherView): boolean {
-    return this.manageableCiphers.some((x) => x.id === c.id);
   }
 }
